Try spectral pooling (with ReLu on real part) (attempt 8)

+++ models/cifar10/cifar10.py
@@ -281,7 +281,7 @@ def inference(images, hyperParam):
         conv = rfft2d(tf.transpose(conv, pre_ffft_transpose))
         width = int(conv.shape[2].value)
         height = int(conv.shape[3].value)
-        conv = conv[:,:,(width/4):(width*3/4),(height/4):(height*3/4)]
+        conv = conv[:,:,int(width/4):int(width*3/4),int(height/4):int(height*3/4)]
         return tf.transpose(irfft2d(layerOut), [0, 2, 3, 1])
     poolfun = spectralPool
